% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookie.R
\name{cookie}
\alias{cookie}
\title{Get yourself a cookie!!}
\usage{
cookie(cookie_type)
}
\arguments{
\item{cookie_type}{options include "chocolate chip", "oatmeal raisin", "sugar", "mystery", "gina"}
}
\value{
a cookie!!
}
\description{
Get yourself a cookie!!
}
\examples{
cookie('chocolate chip')
}
